requirements.txt
alembic==1.13.2
APScheduler==3.10.4
async-timeout==4.0.3
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
msgspec==0.18.6
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.3